+++ src/multicodec/codecs/wrap.clj
@@ -1,21 +1,11 @@
 (ns multicodec.codecs.wrap
-  "Wrapper codec which writes out and reads a header before the value body."
+  "Wrapper codec which writes out and reads a header before the value body. This
+  is equivalent to calling the wrapped codec with the `encode-with-header!` and
+  `decode-with-header!` functions from the core ns."
   (:require
     [multicodec.core :as codec]
     [multicodec.header :as header]))
 
-
-(defn ^:no-doc write-header-encoded!
-  "Convenience function to write a header for the given codec, then use it to
-  write out the encoded value."
-  [codec output header value]
-  (let [header-length (header/write-header! output header)
-        body-length (codec/encode! codec output value)]
-    (+ header-length body-length)))
-
-
-
-;; ## Wrapper Codec
 
 (defrecord WrapperCodec
   [header codec]
@@ -29,7 +19,7 @@
 
   (encode!
     [this output value]
-    (write-header-encoded! codec output header value))
+    (codec/encode-with-header! codec output header value))
 
 
   codec/Decoder
@@ -41,13 +31,7 @@
 
   (decode!
     [this input]
-    (let [header' (header/read-header! input)]
-      (when-not (= header header')
-        (throw (ex-info
-                 (format "The stream header %s did not match expected header %s"
-                         (pr-str header') (pr-str header))
-                 {:expected header, :actual header'})))
-      (codec/decode! codec input))))
+    (codec/decode-with-header! codec input header)))
 
 
 (defn wrap-header